Factor sendEmail function out of calls route
