+++ util/sendDummyCall.js
@@ -27,8 +27,8 @@ const sendDummyCall = async () => {
 
   var options = {
     method: 'POST',
-    // url: 'http://localhost:1337/calls',
-    url: 'https://gfd.gr/calls',
+    url: 'http://localhost:1337/calls',
+    // url: 'https://gfd.gr/calls',
     // url: 'https://testing.gfd.gr/calls',
     qs: dummyCall,
     headers: {
